feat: add new contact list form for the buyer wizard
